requirements.txt
PyYAML==6.0
selenium==4.11.2
beautifulsoup4==4.11.1
prettytable==3.5.0
pytz